% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggPlantmap.heatmap}
\alias{ggPlantmap.heatmap}
\title{ggPlantmap.heatmap}
\usage{
ggPlantmap.heatmap(
  map.quant,
  value.quant = value.quant,
  show.legend = T,
  linewidth = 0.5
)
}
\arguments{
\item{map.quant}{The ggPlantmap tibble with quantitative values.}

\item{value.quant}{The name of the column that contains the quantitative values (Ex: Gene.expression).}

\item{show.legend}{logical. Should a legend for the levels to be included in the plot? If there are too many levels, legends can overwhelm the image. In this case, change to FALSE (Default: TRUE).}

\item{linewdith}{The line width of the tracing lines of your ggplot (Default = 0.5).}
}
\value{
A ggplot generated heatmap of your ggPlantmap with colors that depicts the continuous values of your quantitative data. You can integrate this with a ggplot gradient scale (Example: + scale_fill_gradient())
}
\description{
produce a heatmap from a ggPlantmap that contains quantitative values for specific ROI levels.
}
\examples{
quant.data <- ggPlantmap.merge(map=ggPm.At.seed.developmentalseries,value=ggPm.seed.expression.sample,id.x="ROI.name")
ggPlantmap.heatmap(map.quant=quant.data,value.quant=AT5G47670.expression,linewdith=1,show.legend=T)
}
