% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggPlantmap.to.SVG}
\alias{ggPlantmap.to.SVG}
\title{ggPlantmap.to.SVG}
\usage{
ggPlantmap.to.SVG(
  object,
  group.name = "ROI.name",
  author = "ggPlantmap",
  svg.name = "ggPlantmap.svg"
)
}
\arguments{
\item{object}{The ggPlantmap table object (Ex: ggPm.At.seed.devseries)}

\item{group.name}{Column name (as string) of the column to be used for grouping objects (Default="ROI.name).}

\item{author}{Name of the author (as string) to be attached to the SVG file (Default="ggPlantmap").}

\item{svg.name}{Name of the SVG file (as string) to be saved as (Default="ggPlantmap.svg")}
}
\value{
A SVG file that can be used in any graphic software and it is compatible to create a Plant eFP viewer
}
\description{
Convert the ggPlantmap table into a SVG file that can be used to create a SVG file suitable to create a Plant eFP viewer or worked in graphic software (i.e. Adobe Illustrator, Inkspace, Power Point)
}
\examples{
ggPlantmap.to.SVG(ggPm.At.roottip.longitudinal)
ggPlantmap.to.SVG(ggPm.At.seed.devseries,group.name="Region",author="LJ",svg.name="ggPm.At.seed.devseries.svg")

}
