% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.3weekrosette.topview}
\alias{ggPm.At.3weekrosette.topview}
\title{ggPm.At.roottip.longitudinal}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 806 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 806 rows and 7 columns.
}
\source{
<https://doi.org/10.1186/s13007-019-0417-9>

<https://doi.org/10.1186/s12870-015-0483-8>
}
\usage{
ggPm.At.3weekrosette.topview

ggPm.At.3weekrosette.topview
}
\description{
Longitudinal view of Arabidopsis thaliana root tips.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{root}
  \item{Type}{longitudinal view}
}

Top view of 3.5 week old Arabidopsis thaliana rosettes.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{rosette}
  \item{Type}{top view}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
