% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.flower.diagram}
\alias{ggPm.At.flower.diagram}
\title{ggPm.At.flower.diagram}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 592 rows and 7 columns.
}
\source{
<Taiz, Lincoln, et al. Plant physiology and development. No. Ed. 6. Sinauer Associates Incorporated, 2015.>
}
\usage{
ggPm.At.flower.diagram
}
\description{
Arabidopsis flower diagram.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{flower}
  \item{Type}{diagram}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
