% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.inflorescencestem.crosssection}
\alias{ggPm.At.inflorescencestem.crosssection}
\title{ggPm.At.inflorescencestem.crosssection}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4446 rows and 5 columns.
}
\source{
<https://academic.oup.com/plcell/article/33/2/200/6017181>
}
\usage{
ggPm.At.inflorescencestem.crosssection
}
\description{
Cross-section of Arabidopsis thaliana inflorescence stem.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{inflorescence stem}
  \item{Type}{longitudinal}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
