% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.lateralroot.devseries}
\alias{ggPm.At.lateralroot.devseries}
\title{ggPm.At.lateralroot.devseries}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3415 rows and 7 columns.
}
\source{
<https://doi.org/10.3389/fpls.2019.00206>
}
\usage{
ggPm.At.lateralroot.devseries
}
\description{
Diagram of the development of Arabidopsis thaliana lateral root.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{lateral root}
  \item{Type}{development series}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
