% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.leaf.crosssection}
\alias{ggPm.At.leaf.crosssection}
\title{ggPm.At.leaf.crosssection}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1503 rows and 7 columns.
}
\source{
<https://doi.org/10.1199/tab.0181>
}
\usage{
ggPm.At.leaf.crosssection
}
\description{
Cross-section of Arabidopsis thaliana leaves.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{leaf}
  \item{Type}{cross-section}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
