% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.leaf.topview}
\alias{ggPm.At.leaf.topview}
\title{ggPm.At.leaf.topview}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 935 rows and 7 columns.
}
\source{
<http://doi.org/10.1199/tab.0181>
}
\usage{
ggPm.At.leaf.topview
}
\description{
Top view of Arabidopsis thaliana leaves from left to right in order of emergence from the SAM.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{leaf}
  \item{Type}{top view}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
