% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.leafepidermis.topview}
\alias{ggPm.At.leafepidermis.topview}
\title{ggPm.At.leafepidermis.topview}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2053 rows and 5 columns.
}
\source{
<https://www.nature.com/articles/s41477-021-00906-0>
}
\usage{
ggPm.At.leafepidermis.topview
}
\description{
Top view of Arabidopsis thaliana leaf adaxial epidermis.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{leaf epidermis}
  \item{Type}{top view}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
