% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.rootmatur.crosssection}
\alias{ggPm.At.rootmatur.crosssection}
\title{ggPm.At.rootmatur.crosssection}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1113 rows and 5 columns.
}
\source{
<https://doi.org/10.1016/j.devcel.2022.01.008>
}
\usage{
ggPm.At.rootmatur.crosssection
}
\description{
Cross-section of Arabidopsis thaliana root at the maturation zone.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{root}
  \item{Type}{cross-section}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
