% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.roottip.crosssection}
\alias{ggPm.At.roottip.crosssection}
\title{ggPm.At.roottip.crosssection}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1408 rows and 5 columns.
}
\source{
<https://www.future-science.com/doi/10.2144/000114621>
}
\usage{
ggPm.At.roottip.crosssection
}
\description{
Cross-section of Arabidopsis thaliana root tips.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{root}
  \item{Type}{cross-section}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
