% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.seed.devseries}
\alias{ggPm.At.seed.devseries}
\title{ggPm.At.seed.devseries}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5369 rows and 8 columns.
}
\source{
<https://doi.org/10.1073/pnas.1222061110>
}
\usage{
ggPm.At.seed.devseries
}
\description{
Diagram of the development of Arabidopsis thaliana seed tissues.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{seed}
  \item{Type}{development series}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
