% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.shootapex.longitudinal}
\alias{ggPm.At.shootapex.longitudinal}
\title{ggPm.At.shootapex.longitudinal}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1218 rows and 7 columns.
}
\source{
<https://doi.org/10.1007/s10265-020-01174-3>
}
\usage{
ggPm.At.shootapex.longitudinal
}
\description{
Diagram of the Arabidopsis thaliana shoot apical meristem.
\describe{
  \item{Species}{Arabidopsis thaliana}
  \item{Tissue}{shoot apical meristem}
  \item{Type}{longitudinal}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
