% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.Ms.root.crosssection}
\alias{ggPm.Ms.root.crosssection}
\title{ggPm.Ms.root.crosssection}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2441 rows and 5 columns.
}
\source{
Unpublished image (Leonardo Jo)
}
\usage{
ggPm.Ms.root.crosssection
}
\description{
Cross-section of Medicago sativa (Alfalfa) root.
\describe{
  \item{Species}{Medicago sativa}
  \item{Tissue}{root}
  \item{Type}{cross-section}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
