% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.Sl.root.crosssection}
\alias{ggPm.Sl.root.crosssection}
\title{ggPm.Sl.root.crosssection}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4314 rows and 5 columns.
}
\source{
<https://doi.org/10.1016/j.cell.2021.04.024>
}
\usage{
ggPm.Sl.root.crosssection
}
\description{
Cross-section of Solanum lycopersicum (Tomato) root.
\describe{
  \item{Species}{Solanum lycopersicum}
  \item{Tissue}{root}
  \item{Type}{cross-section}
}
}
\author{
ggPlantmap author: Leonardo Jo (jo.leonardo85@gmail.com)
}
\keyword{datasets}
