% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.summary}
\alias{ggPm.summary}
\title{ggPm.summary}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 9 columns.
}
\usage{
ggPm.summary
}
\description{
Summary of pre-loaded ggPLantmaps (Version Beta.1)
\describe{
  \item{Species}{Species name}
  \item{Tissue}{Type of tissue (Ex: root, leaf, stem)}
  \item{Type}{Type of image (Ex: cross-section, longitudinal view, development series, etc)}
  \item{Description}{Brief description of the ggPlantmap}
  \item{Layers}{Mapped layers in the ggPlantmap (Ex: Cells, Tissues, Stages)}
  \item{Image.Reference}{Link for the refenrece imaged used to create the ggPlantmap}
  \item{Made.by}{Person who made the ggPlantmap}
  \item{Contact.Info}{Author email address}
}
}
\keyword{datasets}
