% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{XML.to.ggPlantmap}
\alias{XML.to.ggPlantmap}
\title{XML.to.ggPlantmap}
\usage{
XML.to.ggPlantmap(data)
}
\arguments{
\item{data}{XML file name (character string; e.g. "arabidopsis.root.crosssection.xml").}
}
\value{
A tibble table with x,y coordinates of individual ROIs (ROI.id) and the name of the ROI (ROI.name)
}
\description{
Open an Icy generated XML file and extract x and y coordinates of ROIs
}
\examples{
XML.to.ggPlantmap("arabidopsis.root.crossection.xml")
}
